Add files via upload
A backend project is ready which is based on Python and machine learning. Various libraries are used in this project  for example numpy, pandas , surprise, matplotlib for plotting the graph and predicting the data which is based on movies dataset. 
An excel file is also there along with the code in which all the dataset is present column by column.